Parse dynamic params on the client
This is part of a larger effort to remove dynamic params from server
responses except in cases where they are needed to render a Server
Component. If a param is not rendered by a Server Component, then it can
omitted from the cache key, and cached responses can be reused across
pages with different param values.

In this step, I've implemented client parsing of the params from
the response headers. This requires sending some additional data in
the response body, such as the regex for the route.

Note: Although the ultimate goal is to remove the dynamic params from
the response body (e.g. FlightRouterState), I have not done so yet this
PR. The rest of the work will be split up into multiple subsequent PRs.