[Segment Cache] Remove params from tree response
Removes the param values from the router tree sent by the server.

This increases the cacheability of the responses because if the response
does not include a certain param in the component data, it can be reused
for all possible values of that param. Previously this was not possible
because even if the components did not reference a param, it was
included in the router tree regardless.