Add app router types
When dealing with app router, there were no types available. Instead,
the user was supposed to define their own types. This is prone to typos
and reduces discoverability of properties.

This change introduces some helper types for dealing with App router.

This also fixes the `React` import by changing it to a star import. The
`react` module has no default export. Instead, Next.js relied on
specific TypeScript configurations that may or may not work in some
situations.