Delete abandoned segment parsing code
Originally, as part of the clientParamParsing feature, I expected
that I would omit param values from the FlightRouterState when sending
data from the server, and add it on the client.

However, I've since realized this is unnecessary because
FlightRouterState is only used during runtime server responses. Not
in static responses, like per-segment prefetches.

Since the goal of clientParamParsing is to omit param values from
static responses, we don't need to change anything
about FlightRouterState.

We may still want to update how params are processed on the server to
more closely match the implementation on the client, but that still
doesn't require any changes to FlightRouterState.

So, this deletes all the code related to adding client-parsed params
to FlightRouterState on the client. None of this was being used
anywhere; it was added by previous PRs in anticipation that it would
eventually be used. But now we don't need it.